Fix clippy cast_precision_loss and float_cmp warnings
- Revert i32 intermediary approach for usize to f64 conversions
- Use direct usize to f64 casts with #[allow] annotations
- Replace exact float comparisons with epsilon-based comparisons
- Use f64::EPSILON for float equality checks in train journey rendering

This is more correct as f64 can represent integers up to 2^53 exactly,
and epsilon comparisons are proper for floating point equality checks.

Reduces warnings from 51 to 36

+++ src/components/error_list.rs
@@ -37,8 +37,9 @@ fn ErrorListPopover(
                                         let conflict_message = conflict.format_message(station1_name, station2_name);
 
                                         let time_fraction = time_to_fraction(conflict.time);
-                                        // Convert via i32 to avoid precision loss (station counts are small)
-                                        let station_position = f64::from(conflict.station1_idx as i32) + conflict.position;
+                                        // Direct usize to f64 conversion is safe for reasonable station counts
+                                        #[allow(clippy::cast_precision_loss)]
+                                        let station_position = conflict.station1_idx as f64 + conflict.position;
 
                                         view! {
                                             <div
